lang-graph/react-intro/pyproject.toml
[project]
name = "react-reasoning-agent"
version = "0.1.0"
description = "基于LangGraph的ReAct(Reasoning and Acting)推理智能体系统"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "autoflake>=2.3.1",
    "langchain>=0.3.26",
    "langgraph>=0.5.2",
    "python-dotenv>=1.0.0",
    "langchain-openai>=0.1.0",
] 